+++ crates/telio-model/src/api_config.rs
@@ -268,6 +268,13 @@ where
     Ok(Some(eps))
 }
 
+/// Turns on the no link detection mechanism
+#[derive(Default, Copy, Clone, Debug, PartialEq, Eq, Deserialize)]
+pub struct FeatureNoLinkDetection {
+    /// Configurable rtt in seconds
+    pub rtt_seconds: Option<u64>,
+}
+
 #[derive(Clone, Debug, Default, PartialEq, Eq, Deserialize)]
 /// Encompasses all of the possible features that can be enabled
 pub struct Features {
@@ -302,6 +309,8 @@ pub struct Features {
     /// Flag to turn on connection reset upon VPN server change for boringtun adapter
     #[serde(default)]
     pub boringtun_reset_connections: FeatureBoringtunResetConns,
+    /// No link detection mechanism
+    pub no_link_detection: Option<FeatureNoLinkDetection>,
 }
 
 impl FeaturePaths {
@@ -453,6 +462,7 @@ mod tests {
         ipv6: true,
         nicknames: true,
         boringtun_reset_connections: FeatureBoringtunResetConns(true),
+        no_link_detection: None,
     });
 
     static EXPECTED_FEATURES_WITHOUT_TEST_ENV: Lazy<Features> = Lazy::new(|| Features {
@@ -494,6 +504,7 @@ mod tests {
         ipv6: false,
         nicknames: false,
         boringtun_reset_connections: FeatureBoringtunResetConns(false),
+        no_link_detection: None,
     });
 
     #[test]
@@ -664,6 +675,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         let empty_qos_features = Features {
@@ -689,6 +701,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         let no_qos_features = Features {
@@ -709,6 +722,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         assert_eq!(from_str::<Features>(full_json).unwrap(), full_features);
@@ -748,6 +762,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         let empty_features = Features {
@@ -764,6 +779,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         assert_eq!(from_str::<Features>(full_json).unwrap(), full_features);
@@ -804,6 +820,7 @@ mod tests {
             ipv6: false,
             nicknames: false,
             boringtun_reset_connections: Default::default(),
+            no_link_detection: None,
         };
 
         assert_eq!(Features::default(), expected_defaults);