+++ changelog.md
@@ -48,6 +48,7 @@
 * LLT-3950: Enable IPv6 for wg-stun.
 * LLT-4502: Implement ICMP, UDP and TCP conntrack tracking for each peer separately
 * LLT-4508: Fetch windows dependencies from local gitlab registry
+* LLT-4545: Add no-link detection mechanism
 
 <br>